+++ src/functionhelper.js
@@ -4,7 +4,10 @@ function FunctionHelper(client) {
   this.client = client;
 }
 
-FunctionHelper.prototype.now = function() {
+FunctionHelper.prototype.now = function(precision) {
+  if (typeof precision === 'number') {
+    return this.client.raw(`CURRENT_TIMESTAMP(${precision})`);
+  }
   return this.client.raw('CURRENT_TIMESTAMP');
 };