+++ src/migrate/Migrator.js
@@ -6,43 +6,42 @@ import mkdirp from 'mkdirp';
 import Promise from 'bluebird';
 import {
   assign,
-  bind,
-  difference,
+  differenceWith,
   each,
   filter,
   get,
-  includes,
   isBoolean,
   isEmpty,
   isUndefined,
-  map,
   max,
   template,
 } from 'lodash';
 import inherits from 'inherits';
+import {
+  getLockTableName,
+  getLockTableNameWithSchema,
+  getTable,
+  getTableName,
+} from './table-resolver';
+import { getSchemaBuilder } from './table-creator';
+import * as migrationListResolver from './migration-list-resolver';
+import FsMigrations, { DEFAULT_LOAD_EXTENSIONS } from './sources/fs-migrations';
 
 function LockError(msg) {
   this.name = 'MigrationLocked';
   this.message = msg;
 }
+
 inherits(LockError, Error);
 
 const CONFIG_DEFAULT = Object.freeze({
   extension: 'js',
-  loadExtensions: [
-    '.co',
-    '.coffee',
-    '.eg',
-    '.iced',
-    '.js',
-    '.litcoffee',
-    '.ls',
-    '.ts',
-  ],
+  loadExtensions: DEFAULT_LOAD_EXTENSIONS,
   tableName: 'knex_migrations',
   schemaName: null,
   directory: './migrations',
   disableTransactions: false,
+  sortDirsSeparately: false,
 });
 
 // The new migration we're performing, typically called from the `knex.migrate`
@@ -51,7 +50,7 @@ const CONFIG_DEFAULT = Object.freeze({
 export default class Migrator {
   constructor(knex) {
     this.knex = knex;
-    this.config = this.setConfig(knex.client.config.migrations);
+    this.config = getMergedConfig(knex.client.config.migrations);
 
     this._activeMigration = {
       fileName: null,
@@ -60,21 +59,26 @@ export default class Migrator {
 
   // Migrators to the latest configuration.
   latest(config) {
-    this.config = this.setConfig(config);
-    return this._migrationData()
-      .tap(validateMigrationList)
+    this.config = getMergedConfig(config, this.config);
+
+    return migrationListResolver
+      .listAllAndCompleted(this.config, this.knex)
+      .tap((value) => validateMigrationList(this.config.migrationSource, value))
       .spread((all, completed) => {
-        const migrations = difference(all, completed);
+        const migrations = getNewMigrations(
+          this.config.migrationSource,
+          all,
+          completed
+        );
 
         const transactionForAll =
           !this.config.disableTransactions &&
           isEmpty(
-            filter(migrations, (name) => {
-              const migration = require(path.join(
-                this._absoluteConfigDir(),
-                name
-              ));
-              return !this._useTransaction(migration);
+            filter(migrations, (migration) => {
+              const migrationContents = this.config.migrationSource.getMigration(
+                migration
+              );
+              return !this._useTransaction(migrationContents);
             })
           );
 
@@ -91,40 +95,48 @@ export default class Migrator {
   // Rollback the last "batch" of migrations that were run.
   rollback(config) {
     return Promise.try(() => {
-      this.config = this.setConfig(config);
-      return this._migrationData()
-        .tap(validateMigrationList)
+      this.config = getMergedConfig(config, this.config);
+
+      return migrationListResolver
+        .listAllAndCompleted(this.config, this.knex)
+        .tap((value) =>
+          validateMigrationList(this.config.migrationSource, value)
+        )
         .then((val) => this._getLastBatch(val))
         .then((migrations) => {
-          return this._runBatch(map(migrations, 'name'), 'down');
+          return this._runBatch(migrations, 'down');
         });
     });
   }
 
   status(config) {
-    this.config = this.setConfig(config);
+    this.config = getMergedConfig(config, this.config);
 
     return Promise.all([
       getTable(this.knex, this.config.tableName, this.config.schemaName).select(
         '*'
       ),
-      this._listAll(),
+      migrationListResolver.listAll(this.config.migrationSource),
     ]).spread((db, code) => db.length - code.length);
   }
 
   // Retrieves and returns the current migration version we're on, as a promise.
   // If no migrations have been run yet, return "none".
   currentVersion(config) {
-    this.config = this.setConfig(config);
-    return this._listCompleted().then((completed) => {
-      const val = max(map(completed, (value) => value.split('_')[0]));
-      return isUndefined(val) ? 'none' : val;
-    });
+    this.config = getMergedConfig(config, this.config);
+
+    return migrationListResolver
+      .listCompleted(this.config.tableName, this.config.schemaName, this.knex)
+      .then((completed) => {
+        const val = max(completed.map((value) => value.split('_')[0]));
+        return isUndefined(val) ? 'none' : val;
+      });
   }
 
   forceFreeMigrationsLock(config) {
-    this.config = this.setConfig(config);
-    const lockTable = this._getLockTableName();
+    this.config = getMergedConfig(config, this.config);
+
+    const lockTable = getLockTableName(this.config.tableName);
     return getSchemaBuilder(this.knex, this.config.schemaName)
       .hasTable(lockTable)
       .then((exist) => exist && this._freeLock());
@@ -132,7 +144,8 @@ export default class Migrator {
 
   // Creates a new migration, with a given name.
   make(name, config) {
-    this.config = this.setConfig(config);
+    this.config = getMergedConfig(config, this.config);
+
     if (!name) {
       return Promise.reject(
         new Error('A name must be specified for the generated migration')
@@ -144,86 +157,22 @@ export default class Migrator {
       .then((val) => this._writeNewMigration(name, val));
   }
 
-  // Lists all available migration versions, as a sorted array.
-  _listAll(config) {
-    this.config = this.setConfig(config);
-    const loadExtensions = this.config.loadExtensions;
-    return Promise.promisify(fs.readdir, { context: fs })(
-      this._absoluteConfigDir()
-    ).then((migrations) => {
-      return filter(migrations, function(value) {
-        const extension = path.extname(value);
-        return includes(loadExtensions, extension);
-      }).sort();
-    });
-  }
-
   // Ensures a folder for the migrations exist, dependent on the migration
   // config settings.
   _ensureFolder() {
-    const dir = this._absoluteConfigDir();
-    return Promise.promisify(fs.stat, { context: fs })(dir).catch(() =>
-      Promise.promisify(mkdirp)(dir)
-    );
-  }
+    const dirs = this._absoluteConfigDirs();
 
-  // Ensures that a proper table has been created, dependent on the migration
-  // config settings.
-  _ensureTable(trx = this.knex) {
-    const { tableName, schemaName } = this.config;
-    const lockTable = this._getLockTableName();
-    const lockTableWithSchema = this._getLockTableNameWithSchema();
-    return getSchemaBuilder(trx, schemaName)
-      .hasTable(tableName)
-      .then(
-        (exists) =>
-          !exists && this._createMigrationTable(tableName, schemaName, trx)
-      )
-      .then(() => getSchemaBuilder(trx, schemaName).hasTable(lockTable))
-      .then(
-        (exists) => !exists && this._createMigrationLockTable(lockTable, trx)
-      )
-      .then(() => getTable(trx, lockTable, this.config.schemaName).select('*'))
-      .then(
-        (data) =>
-          !data.length && trx.into(lockTableWithSchema).insert({ is_locked: 0 })
+    const promises = dirs.map((dir) => {
+      return Promise.promisify(fs.stat, { context: fs })(dir).catch(() =>
+        Promise.promisify(mkdirp)(dir)
       );
-  }
+    });
 
-  _createMigrationTable(tableName, schemaName, trx = this.knex) {
-    return getSchemaBuilder(trx, schemaName).createTable(
-      getTableName(tableName),
-      function(t) {
-        t.increments();
-        t.string('name');
-        t.integer('batch');
-        t.timestamp('migration_time');
-      }
-    );
-  }
-
-  _createMigrationLockTable(tableName, trx = this.knex) {
-    return getSchemaBuilder(trx, this.config.schemaName).createTable(
-      tableName,
-      function(t) {
-        t.increments('index').primary();
-        t.integer('is_locked');
-      }
-    );
-  }
-
-  _getLockTableName() {
-    return this.config.tableName + '_lock';
-  }
-
-  _getLockTableNameWithSchema() {
-    return this.config.schemaName
-      ? this.config.schemaName + '.' + this._getLockTableName()
-      : this._getLockTableName();
+    return Promise.all(promises);
   }
 
   _isLocked(trx) {
-    const tableName = this._getLockTableName();
+    const tableName = getLockTableName(this.config.tableName);
     return getTable(this.knex, tableName, this.config.schemaName)
       .transacting(trx)
       .forUpdate()
@@ -232,7 +181,7 @@ export default class Migrator {
   }
 
   _lockMigrations(trx) {
-    const tableName = this._getLockTableName();
+    const tableName = getLockTableName(this.config.tableName);
     return getTable(this.knex, tableName, this.config.schemaName)
       .transacting(trx)
       .update({ is_locked: 1 });
@@ -254,7 +203,7 @@ export default class Migrator {
   }
 
   _freeLock(trx = this.knex) {
-    const tableName = this._getLockTableName();
+    const tableName = getLockTableName(this.config.tableName);
     return getTable(trx, tableName, this.config.schemaName).update({
       is_locked: 0,
     });
@@ -266,11 +215,27 @@ export default class Migrator {
       this._getLock(trx)
         // When there is a wrapping transaction, some migrations
         // could have been done while waiting for the lock:
-        .then(() => (trx ? this._listCompleted(trx) : []))
-        .then((completed) => (migrations = difference(migrations, completed)))
+        .then(
+          () =>
+            trx
+              ? migrationListResolver.listCompleted(
+                  this.config.tableName,
+                  this.config.schemaName,
+                  trx
+                )
+              : []
+        )
+        .then(
+          (completed) =>
+            (migrations = getNewMigrations(
+              this.config.migrationSource,
+              migrations,
+              completed
+            ))
+        )
         .then(() =>
           Promise.all(
-            map(migrations, bind(this._validateMigrationStructure, this))
+            migrations.map(this._validateMigrationStructure.bind(this))
           )
         )
         .then(() => this._latestBatchNumber(trx))
@@ -294,7 +259,10 @@ export default class Migrator {
               'If you are sure migrations are not running you can release the ' +
                 'lock manually by deleting all the rows from migrations lock ' +
                 'table: ' +
-                this._getLockTableNameWithSchema()
+                getLockTableNameWithSchema(
+                  this.config.tableName,
+                  this.config.schemaName
+                )
             );
           } else {
             if (this._activeMigration.fileName) {
@@ -318,37 +286,23 @@ export default class Migrator {
 
   // Validates some migrations by requiring and checking for an `up` and `down`
   // function.
-  _validateMigrationStructure(name) {
-    const migration = require(path.join(this._absoluteConfigDir(), name));
+  _validateMigrationStructure(migration) {
+    const migrationName = this.config.migrationSource.getMigrationName(
+      migration
+    );
+    const migrationContent = this.config.migrationSource.getMigration(
+      migration
+    );
     if (
-      typeof migration.up !== 'function' ||
-      typeof migration.down !== 'function'
+      typeof migrationContent.up !== 'function' ||
+      typeof migrationContent.down !== 'function'
     ) {
       throw new Error(
-        `Invalid migration: ${name} must have both an up and down function`
+        `Invalid migration: ${migrationName} must have both an up and down function`
       );
     }
-    return name;
-  }
 
-  // Lists all migrations that have been completed for the current db, as an
-  // array.
-  _listCompleted(trx = this.knex) {
-    const { tableName, schemaName } = this.config;
-    return this._ensureTable(trx)
-      .then(() =>
-        trx
-          .from(getTableName(tableName, schemaName))
-          .orderBy('id')
-          .select('name')
-      )
-      .then((migrations) => map(migrations, 'name'));
-  }
-
-  // Gets the migration list from the specified migration directory, as well as
-  // the list of completed migrations to check what should be run.
-  _migrationData() {
-    return Promise.all([this._listAll(), this._listCompleted()]);
+    return migration;
   }
 
   // Generates the stub template for the current migration, returning a compiled
@@ -366,9 +320,12 @@ export default class Migrator {
   // passing any `variables` given in the config to the template.
   _writeNewMigration(name, tmpl) {
     const { config } = this;
-    const dir = this._absoluteConfigDir();
+    const dirs = this._absoluteConfigDirs();
+    const dir = dirs.slice(-1)[0]; // Get last specified directory
+
     if (name[0] === '-') name = name.slice(1);
     const filename = yyyymmddhhmmss() + '_' + name + '.' + config.extension;
+
     return Promise.promisify(fs.writeFile, { context: fs })(
       path.join(dir, filename),
       tmpl(config.variables || {})
@@ -377,13 +334,21 @@ export default class Migrator {
 
   // Get the last batch of migrations, by name, ordered by insert id in reverse
   // order.
-  _getLastBatch() {
+  _getLastBatch([allMigrations]) {
     const { tableName, schemaName } = this.config;
     return getTable(this.knex, tableName, schemaName)
       .where('batch', function(qb) {
         qb.max('batch').from(getTableName(tableName, schemaName));
       })
-      .orderBy('id', 'desc');
+      .orderBy('id', 'desc')
+      .map((migration) => {
+        return allMigrations.find((entry) => {
+          return (
+            this.config.migrationSource.getMigrationName(entry) ===
+            migration.name
+          );
+        });
+      });
   }
 
   // Returns the latest batch number.
@@ -398,8 +363,8 @@ export default class Migrator {
   // Otherwise, rely on the common config. This allows enabling/disabling
   // transaction for a single migration at will, regardless of the common
   // config.
-  _useTransaction(migration, allTransactionsDisabled) {
-    const singleTransactionValue = get(migration, 'config.transaction');
+  _useTransaction(migrationContent, allTransactionsDisabled) {
+    const singleTransactionValue = get(migrationContent, 'config.transaction');
 
     return isBoolean(singleTransactionValue)
       ? singleTransactionValue
@@ -411,24 +376,32 @@ export default class Migrator {
   _waterfallBatch(batchNo, migrations, direction, trx) {
     const trxOrKnex = trx || this.knex;
     const { tableName, schemaName, disableTransactions } = this.config;
-    const directory = this._absoluteConfigDir();
     let current = Promise.bind({ failed: false, failedOn: 0 });
     const log = [];
     each(migrations, (migration) => {
-      const name = migration;
+      const name = this.config.migrationSource.getMigrationName(migration);
       this._activeMigration.fileName = name;
-      migration = require(directory + '/' + name);
+      const migrationContent = this.config.migrationSource.getMigration(
+        migration
+      );
 
       // We're going to run each of the migrations in the current "up".
       current = current
         .then(() => {
-          if (!trx && this._useTransaction(migration, disableTransactions)) {
-            return this._transaction(migration, direction, name);
+          if (
+            !trx &&
+            this._useTransaction(migrationContent, disableTransactions)
+          ) {
+            return this._transaction(migrationContent, direction, name);
           }
-          return warnPromise(migration[direction](trxOrKnex, Promise), name);
+          return warnPromise(
+            this.knex,
+            migrationContent[direction](trxOrKnex, Promise),
+            name
+          );
         })
         .then(() => {
-          log.push(path.join(directory, name));
+          log.push(name);
           if (direction === 'up') {
             return trxOrKnex.into(getTableName(tableName, schemaName)).insert({
               name,
@@ -448,28 +421,63 @@ export default class Migrator {
     return current.thenReturn([batchNo, log]);
   }
 
-  _transaction(migration, direction, name) {
+  _transaction(migrationContent, direction, name) {
     return this.knex.transaction((trx) => {
-      return warnPromise(migration[direction](trx, Promise), name, () => {
-        trx.commit();
-      });
+      return warnPromise(
+        this.knex,
+        migrationContent[direction](trx, Promise),
+        name,
+        () => {
+          trx.commit();
+        }
+      );
     });
   }
 
-  _absoluteConfigDir() {
-    return path.resolve(process.cwd(), this.config.directory);
-  }
-
-  setConfig(config) {
-    return assign({}, CONFIG_DEFAULT, this.config || {}, config);
+  /** Returns  */
+  _absoluteConfigDirs() {
+    const directories = Array.isArray(this.config.directory)
+      ? this.config.directory
+      : [this.config.directory];
+    return directories.map((directory) => {
+      return path.resolve(process.cwd(), directory);
+    });
   }
 }
 
+export function getMergedConfig(config, currentConfig) {
+  // config is the user specified config, mergedConfig has defaults and current config
+  // applied to it.
+  const mergedConfig = assign({}, CONFIG_DEFAULT, currentConfig || {}, config);
+
+  if (
+    config &&
+    // If user specifies any FS related config,
+    // clear existing FsMigrations migrationSource
+    (config.directory ||
+      config.sortDirsSeparately !== undefined ||
+      config.loadExtensions)
+  ) {
+    mergedConfig.migrationSource = null;
+  }
+
+  // If the user has not specified any configs, we need to
+  // default to fs migrations to maintain compatibility
+  if (!mergedConfig.migrationSource) {
+    mergedConfig.migrationSource = new FsMigrations(
+      mergedConfig.directory,
+      mergedConfig.sortDirsSeparately
+    );
+  }
+
+  return mergedConfig;
+}
+
 // Validates that migrations are present in the appropriate directories.
-function validateMigrationList(migrations) {
+function validateMigrationList(migrationSource, migrations) {
   const all = migrations[0];
   const completed = migrations[1];
-  const diff = difference(completed, all);
+  const diff = getMissingMigrations(migrationSource, completed, all);
   if (!isEmpty(diff)) {
     throw new Error(
       `The migration directory is corrupt, the following files are missing: ${diff.join(
@@ -479,9 +487,25 @@ function validateMigrationList(migrations) {
   }
 }
 
-function warnPromise(value, name, fn) {
+function getMissingMigrations(migrationSource, completed, all) {
+  return differenceWith(completed, all, (completedMigration, allMigration) => {
+    return (
+      completedMigration === migrationSource.getMigrationName(allMigration)
+    );
+  });
+}
+
+function getNewMigrations(migrationSource, all, completed) {
+  return differenceWith(all, completed, (allMigration, completedMigration) => {
+    return (
+      completedMigration === migrationSource.getMigrationName(allMigration)
+    );
+  });
+}
+
+function warnPromise(knex, value, name, fn) {
   if (!value || typeof value.then !== 'function') {
-    this.knex.client.logger.warn(`migration ${name} did not return a promise`);
+    knex.client.logger.warn(`migration ${name} did not return a promise`);
     if (fn && typeof fn === 'function') fn();
   }
   return value;
@@ -505,23 +529,4 @@ function yyyymmddhhmmss() {
     padDate(d.getMinutes()) +
     padDate(d.getSeconds())
   );
-}
-
-//Get schema-aware table name
-function getTableName(tableName, schemaName) {
-  return schemaName ? `${schemaName}.${tableName}` : tableName;
-}
-
-//Get schema-aware query builder for a given table and schema name
-function getTable(trxOrKnex, tableName, schemaName) {
-  return schemaName
-    ? trxOrKnex(tableName).withSchema(schemaName)
-    : trxOrKnex(tableName);
-}
-
-//Get schema-aware schema builder for a given schema nam
-function getSchemaBuilder(trxOrKnex, schemaName) {
-  return schemaName
-    ? trxOrKnex.schema.withSchema(schemaName)
-    : trxOrKnex.schema;
 }