+++ test/integration/migrate/index.js
@@ -1,10 +1,11 @@
 /*global after, before, describe, expect, it*/
 'use strict';
 
-var equal = require('assert').equal;
-var path = require('path');
-var rimraf = require('rimraf');
-var Promise = require('bluebird');
+const equal = require('assert').equal;
+const path = require('path');
+const rimraf = require('rimraf');
+const Promise = require('bluebird');
+const testMemoryMigrations = require('./memory-migrations');
 
 module.exports = function(knex) {
   require('rimraf').sync(path.join(__dirname, './migration'));
@@ -31,7 +32,11 @@ module.exports = function(knex) {
       });
     });
 
-    var tables = ['migration_test_1', 'migration_test_2', 'migration_test_2_1'];
+    const tables = [
+      'migration_test_1',
+      'migration_test_2',
+      'migration_test_2_1',
+    ];
 
     describe('knex.migrate.status', function() {
       this.timeout(process.env.KNEX_TEST_TIMEOUT || 30000);
@@ -243,6 +248,45 @@ module.exports = function(knex) {
       });
     });
 
+    describe('knex.migrate.latest - multiple directories', () => {
+      before(() => {
+        return knex.migrate.latest({
+          directory: [
+            'test/integration/migrate/test',
+            'test/integration/migrate/test2',
+          ],
+        });
+      });
+
+      after(() => {
+        return knex.migrate.rollback({
+          directory: [
+            'test/integration/migrate/test',
+            'test/integration/migrate/test2',
+          ],
+        });
+      });
+
+      it('should create tables specified in both directories', () => {
+        // Map the table names to promises that evaluate chai expectations to
+        // confirm that the table exists
+        const expectedTables = [
+          'migration_test_1',
+          'migration_test_2',
+          'migration_test_2_1',
+          'migration_test_3',
+          'migration_test_4',
+          'migration_test_4_1',
+        ];
+
+        return Promise.map(expectedTables, function(table) {
+          return knex.schema.hasTable(table).then(function(exists) {
+            expect(exists).to.equal(true);
+          });
+        });
+      });
+    });
+
     describe('knex.migrate.rollback', function() {
       it('should delete the most recent batch from the migration log', function() {
         return knex.migrate
@@ -253,7 +297,7 @@ module.exports = function(knex) {
             var migrationPath = ['test', 'integration', 'migrate', 'test'].join(
               path.sep
             ); //Test fails on windows if explicitly defining /test/integration/.. ~wubzz
-            expect(log[0]).to.contain(migrationPath);
+            expect(log[0]).to.contain(batchNo);
             return knex('knex_migrations')
               .select('*')
               .then(function(data) {
@@ -323,7 +367,7 @@ module.exports = function(knex) {
         }
       ).then(function(res) {
         // One should fail:
-        var hasLockError =
+        const hasLockError =
           res[0] === 'MigrationLocked' || res[1] === 'MigrationLocked';
         expect(hasLockError).to.equal(true);
 
@@ -495,5 +539,40 @@ module.exports = function(knex) {
         });
       });
     }
+  });
+
+  describe('migrationSource config', function() {
+    const migrationSource = {
+      getMigrations() {
+        console.log('???', Object.keys(testMemoryMigrations).sort());
+        return Promise.resolve(Object.keys(testMemoryMigrations).sort());
+      },
+      getMigrationName(migration) {
+        return migration;
+      },
+      getMigration(migration) {
+        return testMemoryMigrations[migration];
+      },
+    };
+
+    before(() => {
+      return knex.migrate.latest({
+        migrationSource,
+      });
+    });
+
+    after(() => {
+      return knex.migrate.rollback({
+        migrationSource,
+      });
+    });
+
+    it('can accept a custom migration source', function() {
+      return knex.schema
+        .hasColumn('migration_source_test_1', 'name')
+        .then(function(exists) {
+          expect(exists).to.equal(true);
+        });
+    });
   });
 };